+++ .idea/misc.xml
@@ -21,6 +21,7 @@
         <entry key="..\:/Users/ioolv/OneDrive/Do-Drone/app/src/main/res/layout/fragment_camera.xml" value="0.1390625" />
         <entry key="..\:/Users/ioolv/OneDrive/Do-Drone/app/src/main/res/layout/layout_nav_header.xml" value="0.15715579710144928" />
         <entry key="..\:/Users/ioolv/OneDrive/Do-Drone/app/src/main/res/layout/text.xml" value="0.2972222222222222" />
+        <entry key="..\:/Users/ioolv/OneDrive/Do-Drone/app/src/main/res/menu/nav_menu.xml" value="0.14722222222222223" />
       </map>
     </option>
   </component>
