+++ .project
@@ -16,7 +16,7 @@
 	</natures>
 	<filteredResources>
 		<filter>
-			<id>1626930826164</id>
+			<id>0</id>
 			<name></name>
 			<type>30</type>
 			<matcher>
