+++ .settings/org.eclipse.buildship.core.prefs
@@ -5,7 +5,7 @@ connection.gradle.distribution=GRADLE_DISTRIBUTION(WRAPPER)
 connection.project.dir=
 eclipse.preferences.version=1
 gradle.user.home=
-java.home=/Library/Java/JavaVirtualMachines/jdk-14.0.1.jdk/Contents/Home
+java.home=C\:/Program Files/Eclipse Foundation/jdk-11.0.12.7-hotspot
 jvm.arguments=
 offline.mode=false
 override.workspace.settings=true
