+++ app/src/main/java/com/example/dodrone/ModifyUserInfoActivity.java
@@ -1,5 +1,6 @@
 package com.example.dodrone;
 
+import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Intent;
@@ -17,6 +18,13 @@ import com.google.android.gms.auth.api.signin.GoogleSignIn;
 import com.google.android.gms.auth.api.signin.GoogleSignInAccount;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
+import com.google.firebase.database.DataSnapshot;
+import com.google.firebase.database.DatabaseError;
+import com.google.firebase.database.DatabaseReference;
+import com.google.firebase.database.FirebaseDatabase;
+import com.google.firebase.database.ValueEventListener;
+
+import org.jetbrains.annotations.NotNull;
 
 public class ModifyUserInfoActivity extends AppCompatActivity {
     TextView userName, userEmail;
@@ -32,7 +40,7 @@ public class ModifyUserInfoActivity extends AppCompatActivity {
         setContentView(R.layout.activity_modify_user_info);
 
         FirebaseUser currUser = FirebaseAuth.getInstance().getCurrentUser();
-        LoginActivity.User thisUser = new LoginActivity.User();
+        User thisUser = new User();
         thisUser.retrieveUserInfo(currUser, thisUser.listener);
 
         rbClicked =0;
