+++ app/src/main/java/com/example/dodrone/OnGetDataListener.java
@@ -1,0 +1,10 @@
+package com.example.dodrone;
+
+import com.google.firebase.database.DataSnapshot;
+
+public interface OnGetDataListener {
+    void onSuccess(DataSnapshot dataSnapshot);
+    void onStart();
+    void onFailure();
+
+}
