+++ app/src/main/res/layout/activity_do_drone.xml
@@ -115,7 +115,8 @@
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintTop_toTopOf="parent"
             app:srcCompat="@drawable/ic_baseline_menu_24"
-            tools:srcCompat="@drawable/ic_baseline_menu_24" />
+            tools:srcCompat="@drawable/ic_baseline_menu_24"
+            android:fontFamily="@font/hyundai_b"/>
 
 
     </androidx.constraintlayout.widget.ConstraintLayout>
